pyproject.toml
[tool.poetry]
name = "parser-tgc1"
version = "0.1.0"
description = ""
authors = ["Alexey Inkov <alexo982@mail.ru>"]
readme = "README.md"
packages = [{include = "parser_tgc1"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
pytesseract = "^0.3.10"
pillow = "^10.1.0"
pdf2image = "^1.16.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
